qiuzhi-project/qiuzhi-video-maker/requirements.txt
pandas>=2.0.0
openpyxl>=3.0.0
xlrd>=2.0.1
matplotlib>=3.0.0
seaborn>=0.11.0
gtts>=2.2.0
pydub>=0.25.0
moviepy>=2.0.0
Pillow>=9.0.0